backend/requirements.txt
Flask
flask-cors
torch
torchvision
ultralytics
opencv-python
scikit-learn
pandas
joblib
requests
beautifulsoup4
pymongo
numpy
aiohttp
lxml
pillow
google-cloud-vision
selenium
webdriver-manager
easyocr
tensorflow
transformers
seaborn
matplotlib
plotly
bcrypt
PyJWT